cambios en alerta de log out, btn ahora tiene la animacion de loading y email configurado para confirmar pago

+++ client/src/components/CartComponents/Payment.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState, useEffect, useRef } from 'react'
 import { useSelector } from "react-redux";
 import { useContext } from 'react'
 import { CartContext } from './CartContext'
@@ -8,77 +8,78 @@ import { Link, useHistory } from 'react-router-dom'
 import { CardElement, useStripe, useElements } from '@stripe/react-stripe-js'
 import axios from 'axios';
 import firebase from 'firebase'
+import { LoadingButton } from '@mui/lab';
 
 function Payment() {
     const user = useSelector(state => state.user)
-    const { products,setProducts ,setPay} = useContext(CartContext)
+    const { products, setProducts ,setPay} = useContext(CartContext)
+
+    const [loading, setLoading] = useState(false);
 
     const history = useHistory()
 
     const stripe = useStripe()
     const elements = useElements()
 
+    const [errorMsg, setErrorMsg] = useState({
+        value: null,
+        string: ''
+    })
+
     const [error, setError] = useState(null)
+    
     const [disabled, setDisabled] = useState(true)
 
     const [ succeeded, setSucceeded ] = useState(false)
     const [ processing, setProcessing ] = useState("")
-    const [ clientSecret, setClientSecret ] = useState(true)
 
     const [ email, setEmail ] = useState('')
-    const [ phone, setPhone ] = useState('')
+
     ///------------------------------
    
 
     async function handleSubmit(e) {     //
         e.preventDefault()
-      
-        setProcessing(true)
 
-        const { error, paymentMethod } = await stripe.createPaymentMethod({
-            type: 'card',
-            card: elements.getElement(CardElement)
-        });
+        if(email) {
+            setLoading(true)
+            setProcessing(true)
 
-        if(!error) {
-            const { id } = paymentMethod;
+            const { error, paymentMethod } = await stripe.createPaymentMethod({
+                type: 'card',
+                card: elements.getElement(CardElement)
+            });
 
-            try {
-                const { data } = await axios.post('http://localhost:3001/checkout', {
-                    id, 
-                    amount: subTotal * 100, // lo tengo que mandar en centavos  //1 METODO
-                    email,
-                    phone
-                });
+            if(!error) {
+                const { id } = paymentMethod;
 
-                console.log(data)
+                try {
+                    const { data } = await axios.post('http://localhost:3001/checkout', {
+                        id, 
+                        amount: subTotal * 100, // lo tengo que mandar en centavos  //1 METODO
+                        receipt_email: email,
+                    });
+                    
+                    setLoading(false)
+                    setSucceeded(true)
+                    setError(null)
+                    setProcessing(false)    
+                    elements.getElement(CardElement).clear()
+                    await deleteStockFirebase()
+                    //localStorage.clear()
+                    setPay(true)
+                    setProducts([])
+                    alert('Payment successful')
+                    window.localStorage.clear()
+                    history.replace('/success')
+    
+                } catch (error) {
+                    console.log(error)
+                }
 
-                // console.log(data)
-                // db.collection('users')
-                //.doc(user?.id)
-                //.colection('orders')
-                //.doc(data.id)
-                //.set({
-                //  basket: data.basket,
-                // amount: data.amount,
-                // created: data.created  
-                //})
-
-                setSucceeded(true)
-                setError(null)
-                setProcessing(false)    
-                elements.getElement(CardElement).clear()
-                await deleteStockFirebase()
-                  //localStorage.clear()
-                  setPay(true)
-                  setProducts([])
-                alert('Payment successful')
-                window.localStorage.clear()
-                history.replace('/success')
- 
-            } catch (error) {
-                console.log(error)
             }
+        } else {
+            alert('Form not complete')
         }
     }
 
@@ -106,15 +107,25 @@ function Payment() {
                 .then(()=>{
                     console.log("flight completed")
                 })
-                
             }
-        })
-        
+        }) 
     }
 
     function handleChange(e) {
-        setDisabled(e.empty)                           //1 METODO
+        setLoading(false)
+        setDisabled(e.empty)
         setError(e.error ? e.error.message : "")
+        if(e.error) {
+            setErrorMsg({
+                value: true,
+                string: e.error.message
+            })            
+        } else {
+            setErrorMsg({
+                value: false,
+                string: ''
+            })    
+        }
     }
 
     const [subTotal, setSubTotal] = useState()
@@ -153,33 +164,36 @@ function Payment() {
 
         {/* PAYMENT METHOD */}
         <h1>Payment Method</h1>
-        <form className={css.form_container} onSubmit={handleSubmit}>
+        <form className={css.form_container}>
             <label>Email</label>
             <input 
             type="text" 
             value={email} 
             name='email'
             onChange={e => setEmail(e.target.value)}
+            required
             />
 
-            <label>Phone</label>
-            <input 
-            type="text" 
-            value={phone} 
-            name='phone'
-            onChange={e => setPhone(e.target.value)}
-            />
             <br />
             <CardElement onChange={handleChange}/>
             <br />
+
             <div>
             <h5>Order Total:</h5>{ subTotal && <span>${subTotal}</span>}
             </div>
-            <button disabled={processing || disabled || succeeded}>
-                <span>{processing ? <p>Processing</p> : 'Buy now'}</span>
-            </button>
 
-            {error && <span>{error}</span>}
+            <br />
+
+            <LoadingButton
+                onClick={handleSubmit}
+                endIcon='✔'
+                loading={loading}
+                loadingPosition="end"
+                variant="contained"
+                disabled={processing || disabled || succeeded || errorMsg.value}
+                ><span>{loading ? <p>Processing</p> : 'Buy now'}</span></LoadingButton>
+
+            {errorMsg.string && <span>{errorMsg.string}</span>}
         </form>
     </div>
   )
